# backend/openapi.yaml
openapi: 3.0.3
info:
  title: MultiSubs BaaS API
  description: |
    Multi-tenant Banking-as-a-Service backend for card authorization,
    settlement, yield management, and on-chain treasury operations.
  version: 0.1.0
  contact:
    name: MultiSubs Team
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.multisubs.io
    description: Production
security:
  - ApiKeyAuth: []
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        SHA-256 hashed API key issued per tenant. Prefix: `msk_`.
        In development, `X-Test-Tenant-Id` header may bypass key lookup.
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: "Internal server error"
...
    DashboardResponse:
      type: object
      properties:
        totalTenants:
          type: integer
        tenantsByStatus:
          type: object
          properties:
            active:
              type: integer
            suspended:
              type: integer
            disabled:
              type: integer
        totalUsers:
          type: integer
        totalSubAccounts:
          type: integer
        totalTransactionsToday:
          type: integer
        totalVolumeToday:
          type: string
          description: "BigInt as string (cents)"
    TreasuryResponse:
      type: object
      properties:
        addresses:
          type: object
          properties:
            m1Treasury:
              type: string
              nullable: true
            platformIssuerSafe:
              type: string
              nullable: true
            usdc:
              type: string
              nullable: true
            morphoVault:
              type: string
              nullable: true
            treasuryVault:
              type: string
              nullable: true
        lastSweepTimestamp:
          type: string
          format: date-time
          nullable: true
        lastYieldSnapshotTimestamp:
          type: string
          format: date-time
          nullable: true
    SettlementQueueResponse:
      type: object
      properties:
        pendingCount:
          type: integer
        failedCount:
          type: integer
        averageSettlementTimeMs:
          type: integer
          nullable: true
        recentFailures:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              lithicTxToken:
                type: string
              amount:
                type: string
              errorMessage:
                type: string
                nullable: true
              retryCount:
                type: integer
              createdAt:
                type: string
                format: date-time
              updatedAt:
                type: string
                format: date-time
...
paths:
...
  /v1/admin/dashboard:
    get:
      tags: [Admin]
      summary: Platform dashboard overview
      description: |
        Multi-tenant overview including total tenants, users, sub-accounts,
        and today's transaction volume. Admin only.
      operationId: getAdminDashboard
      responses:
        "200":
          description: Dashboard metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden (admin only)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/admin/treasury:
    get:
      tags: [Admin]
      summary: Treasury health
      description: |
        Returns M1 Treasury configuration addresses and timestamps for the
        last sweep and yield snapshot. Admin only.
      operationId: getAdminTreasury
      responses:
        "200":
          description: Treasury info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TreasuryResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden (admin only)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/admin/settlement-queue:
    get:
      tags: [Admin]
      summary: Settlement queue status
      description: |
        Returns pending/failed settlement counts, average settlement time,
        and the 10 most recent settlement failures. Admin only.
      operationId: getAdminSettlementQueue
      responses:
        "200":
          description: Settlement queue status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettlementQueueResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden (admin only)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...